+++ samplesheets/models.py
@@ -614,7 +614,7 @@ class Assay(BaseSampleSheet):
                 kwargs={'project': self.get_project().sodar_uuid},
             )
             + f'#/study/{self.study.sodar_uuid}'
-            + f'#/assay/{self.sodar_uuid}'
+            + f'/assay/{self.sodar_uuid}'
         )