+++ samplesheets/tests/test_models.py
@@ -844,7 +844,7 @@ class TestAssay(SamplesheetsModelTestBase):
                 kwargs={'project': self.project.sodar_uuid},
             )
             + f'#/study/{self.study.sodar_uuid}'
-            + f'#/assay/{self.assay.sodar_uuid}'
+            + f'/assay/{self.assay.sodar_uuid}'
         )
         self.assertEqual(self.assay.get_url(), expected)