+++ CHANGELOG.rst
@@ -16,7 +16,7 @@ Fixed
     - Missing input validations in data object queries (#2564)
 - **Samplesheets**
     - iRODS file search fails with upper case search terms (#2560)
-    - Vue3app initial routing fails with ``Assay.get_url()`` (#2559)
+    - Vue3app initial routing fails with ``Assay.get_url()`` (#2559, #2565)
     - Vue3app study tab inactive after study or assay URL init (#2561)